% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-vector.R
\name{add_flatgeobuf}
\alias{add_flatgeobuf}
\title{Add a FlatGeobuf layer}
\usage{
add_flatgeobuf(map, data, name = "FlatGeobuf", ...)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{data}{A dataset URL, a local file path, or an \code{sf} object.}

\item{name}{Layer name.}

\item{...}{Additional style overrides given as named arguments, merged into
\code{style}. \code{add_geojson(map, data, fillColor = "red")} and
\code{add_geojson(map, data, style = list(fillColor = "red"))} are equivalent.}
}
\value{
The modified widget.
}
\description{
Add a FlatGeobuf layer
}
\examples{
geolibre() |> add_flatgeobuf("https://example.com/data.fgb")
}
