% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-vector.R
\name{add_geoparquet}
\alias{add_geoparquet}
\title{Add a GeoParquet layer}
\usage{
add_geoparquet(map, data, name = "GeoParquet", ...)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{data}{A dataset URL, a local file path, or an \code{sf} object.}

\item{name}{Layer name.}

\item{...}{Additional style overrides given as named arguments, merged into
\code{style}. \code{add_geojson(map, data, fillColor = "red")} and
\code{add_geojson(map, data, style = list(fillColor = "red"))} are equivalent.}
}
\value{
The modified widget.
}
\description{
Add a GeoParquet layer
}
\examples{
geolibre() |> add_geoparquet("https://example.com/data.parquet")
}
