% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{builtin_legend_names}
\alias{builtin_legend_names}
\title{Built-in legend preset names}
\usage{
builtin_legend_names()
}
\value{
A character vector of preset names.
}
\description{
The preset names accepted by the \code{builtin} argument of \code{\link[=add_legend]{add_legend()}}.
}
\examples{
builtin_legend_names()
}
