% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-manage.R
\name{clear_layers}
\alias{clear_layers}
\title{Remove every layer}
\usage{
clear_layers(map)
}
\arguments{
\item{map}{A GeoLibre widget.}
}
\value{
The modified widget.
}
\description{
Remove every layer
}
\examples{
map <- geolibre() |> add_marker(-77, 39) |> clear_layers()
stopifnot(length(map$x$project$layers) == 0L)
}
