% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{geolibreOutput}
\alias{geolibreOutput}
\alias{renderGeolibre}
\title{Shiny bindings for GeoLibre}
\usage{
geolibreOutput(outputId, width = "100\%", height = "700px")

renderGeolibre(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Widget dimensions.}

\item{expr}{An expression that generates a GeoLibre widget.}

\item{env}{Environment in which to evaluate \code{expr}.}

\item{quoted}{Whether \code{expr} is quoted.}
}
\description{
Shiny bindings for GeoLibre
}
