% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-manage.R
\name{layer_names}
\alias{layer_names}
\title{Layer names in draw order}
\usage{
layer_names(x)
}
\arguments{
\item{x}{A GeoLibre widget or a project list.}
}
\value{
A character vector of layer names, bottom layer first.
}
\description{
Layer names in draw order
}
\examples{
map <- geolibre() |> add_marker(-77, 39, name = "Pin")
layer_names(map)
}
