% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-manage.R
\name{remove_layer}
\alias{remove_layer}
\title{Remove a layer}
\usage{
remove_layer(map, layer)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{layer}{A layer id or layer name.}
}
\value{
The modified widget.
}
\description{
Remove a layer
}
\examples{
map <- geolibre() |>
  add_marker(-77, 39, name = "Pin") |>
  remove_layer("Pin")
stopifnot(length(map$x$project$layers) == 0L)
}
