% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-manage.R
\name{rename_layer}
\alias{rename_layer}
\title{Rename a layer}
\usage{
rename_layer(map, layer, name)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{layer}{A layer id or layer name.}

\item{name}{The new display name. Surrounding whitespace is stripped.}
}
\value{
The modified widget.
}
\description{
Rename a layer
}
\examples{
map <- geolibre() |>
  add_marker(-77, 39, name = "Pin") |>
  rename_layer("Pin", "Capital")
stopifnot(layer_names(map) == "Capital")
}
