% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{set_basemap}
\alias{set_basemap}
\alias{add_basemap}
\title{Set the background basemap}
\usage{
set_basemap(map, basemap)

add_basemap(map, basemap)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{basemap}{A basemap name from \code{\link[=basemaps]{basemaps()}} or a MapLibre style JSON URL.}
}
\value{
The modified widget.
}
\description{
Set the background basemap
}
\examples{
map <- geolibre() |> set_basemap("dark")
map$x$project$basemapStyleUrl
}
\seealso{
\code{\link[=basemaps]{basemaps()}}, \code{\link[=add_tile_layer]{add_tile_layer()}} for raster basemaps such as
OpenStreetMap.
}
