% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{set_bearing}
\alias{set_bearing}
\title{Set the camera bearing}
\usage{
set_bearing(map, bearing)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{bearing}{Clockwise rotation in degrees.}
}
\value{
The modified widget.
}
\description{
Set the camera bearing
}
\examples{
map <- geolibre() |> set_bearing(45)
stopifnot(map$x$project$mapView$bearing == 45)
}
