% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{set_center}
\alias{set_center}
\title{Center the map}
\usage{
set_center(map, lng, lat, zoom = NULL)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{lng}{Longitude of the new center.}

\item{lat}{Latitude of the new center.}

\item{zoom}{Optional zoom level.}
}
\value{
The modified widget.
}
\description{
Center the map
}
\examples{
map <- geolibre() |> set_center(-77.0369, 38.9072, zoom = 11)
stopifnot(map$x$project$mapView$zoom == 11)
}
