% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-manage.R
\name{set_layer_opacity}
\alias{set_layer_opacity}
\title{Set a layer's opacity}
\usage{
set_layer_opacity(map, layer, opacity)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{layer}{A layer id or layer name.}

\item{opacity}{Opacity from zero to one.}
}
\value{
The modified widget.
}
\description{
Set a layer's opacity
}
\examples{
map <- geolibre() |>
  add_marker(-77, 39, name = "Pin") |>
  set_layer_opacity("Pin", 0.4)
stopifnot(map$x$project$layers[[1]]$opacity == 0.4)
}
