% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-manage.R
\name{set_layer_visibility}
\alias{set_layer_visibility}
\alias{show_layer}
\alias{hide_layer}
\title{Show or hide a layer}
\usage{
set_layer_visibility(map, layer, visible = TRUE)

show_layer(map, layer)

hide_layer(map, layer)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{layer}{A layer id or layer name.}

\item{visible}{\code{TRUE} to show the layer, \code{FALSE} to hide it.}
}
\value{
The modified widget.
}
\description{
Show or hide a layer
}
\examples{
map <- geolibre() |>
  add_marker(-77, 39, name = "Pin") |>
  set_layer_visibility("Pin", FALSE)
stopifnot(isFALSE(map$x$project$layers[[1]]$visible))
}
