% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{set_pitch}
\alias{set_pitch}
\title{Set the camera pitch}
\usage{
set_pitch(map, pitch)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{pitch}{Tilt in degrees, clamped to between 0 and 85.}
}
\value{
The modified widget.
}
\description{
Set the camera pitch
}
\examples{
map <- geolibre() |> set_pitch(60)
stopifnot(map$x$project$mapView$pitch == 60)
}
