% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{set_project_name}
\alias{set_project_name}
\title{Set the project name}
\usage{
set_project_name(map, name)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{name}{The project's display name, as shown in the application and
stored in the project file.}
}
\value{
The modified widget.
}
\description{
Set the project name
}
\examples{
map <- geolibre() |> set_project_name("Chesapeake Bay")
map$x$project$name
}
