% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{set_zoom}
\alias{set_zoom}
\title{Set the map zoom}
\usage{
set_zoom(map, zoom)
}
\arguments{
\item{map}{A GeoLibre widget.}

\item{zoom}{Zoom level, clamped to between 0 and 24.}
}
\value{
The modified widget.
}
\description{
Set the map zoom
}
\examples{
map <- geolibre() |> set_zoom(6)
stopifnot(map$x$project$mapView$zoom == 6)
}
