% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{update_geolibre}
\alias{update_geolibre}
\title{Replace the project displayed by a GeoLibre Shiny widget}
\usage{
update_geolibre(proxy, map)
}
\arguments{
\item{proxy}{A GeoLibre proxy created by \code{geolibre_proxy()}.}

\item{map}{A GeoLibre widget or project list.}
}
\value{
The proxy, invisibly.
}
\description{
Replace the project displayed by a GeoLibre Shiny widget
}
