% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{geolibre_proxy}
\alias{geolibre_proxy}
\title{Create a GeoLibre Shiny proxy}
\usage{
geolibre_proxy(outputId, session = NULL)
}
\arguments{
\item{outputId}{ID of an existing GeoLibre widget.}

\item{session}{A Shiny session.}
}
\value{
A proxy object.
}
\description{
Create a GeoLibre Shiny proxy
}
\examples{
if (interactive() && requireNamespace("shiny", quietly = TRUE)) {
  proxy <- geolibre_proxy("map")
}
}
